Do not rely on GITHUB_ENV for REQ_ID and RESULTS_URL
The GITHUB_ENV file solution doesn't work for workflows that
have 2 or more invocations of this action, as only first VAR=VALUE
instruction is respected, which resulted in bad urls in the result
comment for each testing farm request but the first one.
Relying on step's output variables should fix the issue.