action.yml: add sha input
When using this input it is no longer needed to clone repo and checkout
the PR before calling the `testing-farm-as-a-github-action`, when user
wants updates to the PR status.

If the SHA is not provided, it is obtained same as before, with $(git
rev-parse HEAD). If this fails the action is halted.

SHA obtaining is enabled only when the `update_pull_request_status`
input is enabled.