print github summary
In the GitHub Actions tab a summary of ran job would be displayed if
enabled in inputs. Compose, arch, infra state and testresults are
available.

The github summary is enabled by default. User can disable it by setting
the create_github_summary input to false;

Moreover the user can insert 'key=value' value into this input. This way some
additional information can be propagated to the
GITHUB_STEP_SUMMARY variable as a column in a table.