set default value of status input to false
The update_pull_request_status input is now set to false by default.

This change is not backwards compatible.
Users, who wish to have the status in the PR displayed need to specify
this input explicitly in their yaml configuration file.

Reason for this change is, that there are more possible triggers for
TFaGA than Pull Request and Issue Comment and in those this setting is
not applicable when set to true.o

The tests have been adjusted so that only minimal test tests the
configuration when the update_pull_request_status is disabled.
The other tests test non-default behaviour when the PR update is
enabled, as they did test until now.