fix: don't cancel TF request when GHA Runner cancel the job
Public GitHub Actions Runner has limit of 6 hours for job execution
time.

> **Job execution time** - Each job in a workflow can run for up to 6
hours of execution time. If a job reaches this limit, the job is
terminated and fails to complete.

https://docs.github.com/en/actions/learn-github-actions/usage-limits-billing-and-administration#usage-limits

Let's try to detect this instances and don't cancel the TF requests to
allow users to inspect the results later.