feat: add support for issue metadata and dynamic summary
It replaces the old comment logic with the new one that uses issue metadata
to story helper data. It allows us to create only one comment with a summary
of all tests and their results which is dynamically updated.

Also enable issue comments for our Testing Farm GitHub Action repository.